refactor(mcp): split memory.ts tool bundle into memory/search/bulk (A17.1)
The 576-line mcp/tools/memory.ts bundled seven tools across three concerns.
Split by cohesion:
  - memory.ts  — memory CRUD (remember, update_memory, delete_memory)
  - search.ts  — cross-type retrieval (recall, find_similar, query)
  - bulk.ts    — bulk_write (cross-type batch writer over brain/bulk.ts)

The tool registry (tools/index.ts) is unchanged in content and order — only the
import sources moved — so tools/list and every derived gate stay identical. The
move also drops ten dead imports the bundle had accumulated since bulk_write was
extracted to brain/bulk.ts (createChrono/upsertEntity/upsertEdge/validate* were
import-only). Largest MCP tool file: 576 -> 214 lines. Internal, no behavior
change; covered by the existing mcp-tools suite (94/94).
