+++ server/src/mcp/tools/bulk.ts
@@ -1,0 +1,118 @@
+/**
+ * MCP `bulk_write` tool — batch upsert across knowledge types in one call.
+ *
+ * A cross-type writer (memories + entities + edges + chrono), so it lives in its own file rather
+ * than with memory CRUD. The actual batch logic is the shared `bulkWrite()` in `brain/bulk.ts`
+ * (one source of truth with the REST `POST /bulk` route); this tool only coerces input, fires the
+ * single `bulk.write` summary webhook, and shapes the response.
+ */
+
+import type { ToolHandler, ToolContext, ToolResult, ToolSchemas } from './types.js';
+import { bulkWrite, bulkWriteTotal } from '../../brain/bulk.js';
+import { resolveWriteTarget } from '../../spaces/proxy.js';
+import { emitWebhookEvent } from '../../webhooks/dispatcher.js';
+
+export const bulk_writeTool: ToolHandler = {
+  name: 'bulk_write',
+  description: 'Batch upsert memories, entities, edges, and/or chrono entries in a single call. Processing order: memories → entities → edges → chrono, so edges referencing newly created entities within the same batch resolve correctly. Each array is optional and capped at 500 entries. Per-item validation errors are reported in `errors` without aborting the rest of the batch.',
+  mutating: true,
+  spaceRequired: true,
+  inputSchema: (s: ToolSchemas) => ({
+          type: 'object',
+          properties: {
+            space: s.requiredSpace,
+            memories: {
+              type: 'array',
+              description: 'Memory entries to insert. Same fields as the `remember` tool.',
+              items: {
+                type: 'object',
+                properties: {
+                  fact:        { type: 'string', description: 'The fact or memory to store.' },
+                  tags:        { type: 'array', items: { type: 'string' }, description: 'Categorisation tags.' },
+                  entityIds:   { type: 'array', items: { type: 'string' }, description: 'Related entity IDs.' },
+                  description: { type: 'string', description: 'Optional prose context.' },
+                  type:        { type: 'string', description: 'Optional memory type — selects the per-type schema used to validate `properties`.' },
+                  properties:  { type: 'object', additionalProperties: { oneOf: [{ type: 'string' }, { type: 'number' }, { type: 'boolean' }] } },
+                },
+                required: ['fact'],
+              },
+            },
+            entities: {
+              type: 'array',
+              description: 'Entity entries to upsert. Same fields as the `upsert_entity` tool.',
+              items: {
+                type: 'object',
+                properties: {
+                  id:          { type: 'string', description: 'Optional UUID v4 — if provided, updates the entity with this ID (or inserts with this ID). If omitted, a new entity is always inserted.' },
+                  name:        { type: 'string', description: 'Entity name.' },
+                  type:        { type: 'string', description: 'Entity type.' },
+                  tags:        { type: 'array', items: { type: 'string' } },
+                  description: { type: 'string' },
+                  properties:  { type: 'object', additionalProperties: { oneOf: [{ type: 'string' }, { type: 'number' }, { type: 'boolean' }] } },
+                },
+                required: ['name', 'type'],
+              },
+            },
+            edges: {
+              type: 'array',
+              description: 'Edge entries to upsert. Same fields as the `upsert_edge` tool.',
+              items: {
+                type: 'object',
+                properties: {
+                  from:        { type: 'string', description: 'Source entity ID.' },
+                  to:          { type: 'string', description: 'Target entity ID.' },
+                  label:       { type: 'string', description: 'Relationship label.' },
+                  type:        { type: 'string' },
+                  weight:      { type: 'number' },
+                  description: { type: 'string' },
+                  tags:        { type: 'array', items: { type: 'string' } },
+                  properties:  { type: 'object', additionalProperties: { oneOf: [{ type: 'string' }, { type: 'number' }, { type: 'boolean' }] } },
+                },
+                required: ['from', 'to', 'label'],
+              },
+            },
+            chrono: {
+              type: 'array',
+              description: 'Chrono entries to insert. Same fields as the `create_chrono` tool.',
+              items: {
+                type: 'object',
+                properties: {
+                  title:       { type: 'string' },
+                  type:        { type: 'string', description: 'Entry type (e.g. event, deadline, plan, prediction, milestone, or a custom type defined in the space schema).' },
+                  startsAt:    { type: 'string', description: 'ISO 8601 start date/time.' },
+                  endsAt:      { type: 'string' },
+                  status:      { type: 'string', enum: ['upcoming', 'active', 'completed', 'overdue', 'cancelled'] },
+                  confidence:  { type: 'number' },
+                  description: { type: 'string' },
+                  tags:        { type: 'array', items: { type: 'string' } },
+                  entityIds:   { type: 'array', items: { type: 'string' } },
+                  memoryIds:   { type: 'array', items: { type: 'string' } },
+                  properties:  { type: 'object', additionalProperties: { oneOf: [{ type: 'string' }, { type: 'number' }, { type: 'boolean' }] } },
+                },
+                required: ['title', 'type', 'startsAt'],
+              },
+            },
+            targetSpace: { type: 'string', description: 'Required for proxy spaces: the member space to write to.' },
+          },
+          required: ['space'],
+        }),
+  async handle(ctx: ToolContext): Promise<ToolResult> {
+    const { args: a, callSpace } = ctx;
+    const wt = resolveWriteTarget(callSpace, a['targetSpace'] as string | undefined);
+    if (!wt.ok) throw new Error(wt.error);
+    const ts = wt.target;
+
+    const result = await bulkWrite(ts, {
+      memories: a['memories'], entities: a['entities'], edges: a['edges'], chrono: a['chrono'],
+    });
+    if (bulkWriteTotal(result) > 0) {
+      // Bulk suppresses per-item webhooks; emit ONE summary a workflow can inspect.
+      emitWebhookEvent({ event: 'bulk.write', spaceId: ts, entry: { inserted: result.inserted, updated: result.updated, errorCount: result.errors.length }, ...(ctx.actor ?? {}) });
+    }
+    const summary = `bulk_write complete — inserted: ${JSON.stringify(result.inserted)}, updated: ${JSON.stringify(result.updated)}, errors: ${result.errors.length}`;
+    return {
+      content: [{ type: 'text' as const, text: summary + (result.errors.length > 0 ? '\n' + JSON.stringify(result.errors, null, 2) : '') }],
+      isError: false,
+    };
+  },
+};
